Almost finished bad-input-handling in Blackjack Oving5
Missing to handle integeres that are not 1 or 2 in "if you want to play
more", but cannot be bothered anymore right now. The code is far from
optimal as it is, but I get the point. The error handlign was not part
of the task anyway. (could have made a check-variable, as I did in
BlackJack.cpp, to fix the last input-handling in main). Probably not the
best solution anyway.